+++ mongod/adapter.go
@@ -2,9 +2,8 @@ package mongod
 
 import (
 	"context"
-	"fmt"
-	"log"
 
+	"github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"go.mongodb.org/mongo-driver/x/mongo/driver/connstring"
@@ -21,16 +20,16 @@ func ConnectToDatabase(mongoUrl string, flags map[string]interface{}) (*mongo.Cl
 	// connect to the MongoDB server
 	client, err := mongo.Connect(context.TODO(), clientOptions)
 	if err != nil {
-		log.Fatal(err)
+		logrus.Fatal(err)
 	}
 
 	// check if we were able to connect to the server
 	err = client.Ping(context.TODO(), nil)
 	if err != nil {
-		log.Fatal(err)
+		logrus.Fatal(err)
 	}
 
-	fmt.Println("Connected to MongoDB!")
+	logrus.Infoln("Connected to MongoDB!")
 	if flags["migration"].(bool) {
 		RunMigrations(client, uriOptions.Database, flags)
 	}